Renaming command_runnner to fix spelling issue. Fixing running commands locally and fixing running sudo commands in command_runner.py.
