+++ scheduler_failover_controller/command_runner/command_runner.py
@@ -14,12 +14,12 @@ class CommandRunner:
     # returns: is_successful, output
     def run_command(self, host, base_command):
         self.logger.debug("Running Command: " + str(base_command))
-        # todo get _run_local_command(base_command) working
-        if False:  # host == self.local_hostname or host in self.HOST_LIST_TO_RUN_LOCAL:
+        if host == self.local_hostname or host in self.HOST_LIST_TO_RUN_LOCAL:
             return self._run_local_command(base_command)
         else:
             return self._run_ssh_command(host, base_command)
 
+    # This will start the process up as a child process. Meaning if the scheduler_failover_controller fails the child process will fail as well.
     def _run_local_command(self, base_command):
         self.logger.debug("Running command as Local command")
         output = os.popen(base_command).read()
@@ -31,23 +31,19 @@ class CommandRunner:
     def _run_ssh_command(self, host, base_command):
         self.logger.debug("Running command as SSH command")
         if base_command.startswith("sudo"):
-            command_split = ["ssh", "-t", host, '"' + base_command + '"']
-            return self._run_split_command(
-                command_split=command_split,
-                shell=True
-            )
+            command_split = ["ssh", "-t", host, base_command]
         else:
             command_split = ["ssh", host, base_command]
-            return self._run_split_command(
-                command_split=command_split
-            )
+        return self._run_split_command(
+            command_split=command_split
+        )
 
-    def _run_split_command(self, command_split, shell=False):
-        self.logger.debug("Running command_split: " + str(command_split) + ", shell: " + str(shell))
+    def _run_split_command(self, command_split):
+        self.logger.debug("Running command_split: " + str(command_split))
         is_successful = True
         output = []
         try:
-            process = subprocess.Popen(command_split, shell=shell, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            process = subprocess.Popen(command_split, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             process.wait()
             if process.stderr is not None:
                 stderr_output = process.stderr.readlines()
@@ -61,5 +57,3 @@ class CommandRunner:
             output = str(e)
         self.logger.debug("Run Command output: " + str(output))
         return is_successful, output
-
-
